cloud-functions/gcs-to-bq/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
pandas==2.*
pandas-gbq==0.19.*
